separate camera orientations to new module, prevent circ import
