+++ napari/utils/camera_orientations.py
@@ -1,0 +1,37 @@
+from enum import auto
+from typing import Literal
+
+from napari.utils.misc import StringEnum
+
+
+class VerticalAxisOrientation(StringEnum):
+    UP = auto()
+    DOWN = auto()
+
+
+class HorizontalAxisOrientation(StringEnum):
+    LEFT = auto()
+    RIGHT = auto()
+
+
+class DepthAxisOrientation(StringEnum):
+    AWAY = auto()
+    TOWARDS = auto()
+
+
+class Handedness(StringEnum):
+    RIGHT = auto()
+    LEFT = auto()
+
+
+VerticalAxisOrientationStr = Literal['up', 'down']
+HorizontalAxisOrientationStr = Literal['left', 'right']
+DepthAxisOrientationStr = Literal['away', 'towards']
+
+
+DEFAULT_ORIENTATION_TYPED = (
+    DepthAxisOrientation.TOWARDS,
+    VerticalAxisOrientation.DOWN,
+    HorizontalAxisOrientation.RIGHT,
+)
+DEFAULT_ORIENTATION = tuple(map(str, DEFAULT_ORIENTATION_TYPED))
